refactor: Refactor frontend services and types
- Add SuccessResponse wrapper type to match backend response format
- Update request() helper to auto-unwrap SuccessResponse.data
- Make Category.description and Category.icon nullable (optional)
- Make Product.category nullable (optional)
- Remove OrderItem.product field
- Add new types: OrderUpdate, ProductsOnShop, ProductShop, CampaignCreate, CampaignUpdate, CampaignResponse, MidtransConfigResponse
- Split types/api.ts into separate files: user.type.ts, product.type.ts, shop.type.ts, order.type.ts, payment.type.ts, category.type.ts, campaign.type.ts, common.type.ts
- Split services/api.ts into separate files: user.service.ts, product.service.ts, shop.service.ts, order.service.ts, payment.service.ts, category.service.ts, campaign.service.ts, http.ts
- Create barrel files (index.ts) in both types/ and services/ for easier imports
- Add new service functions: toggleProductPublish, fetchUserProducts, fetchShopByUsername, updateOrderStatus, cancelOrder, and campaign endpoints (create, listActive, update, toggleActive, delete)
- Rename getCategory() to listCategories() for consistency
- Update all imports across the project to use new barrel file structure
- Fix nullable field handling in ProductDetailPage breadcrumbs
- Fix null checking for category.icon in CategoryList
- Remove unused imports

Closes #10
