+++ frontend/src/components/navigation/BottomAppBar.tsx
@@ -1,4 +1,4 @@
-import { AppBar, Button, IconButton, Stack, Toolbar, Typography } from "@mui/material"
+import { AppBar, Button, Stack, Toolbar, Typography } from "@mui/material"
 import { 
     HomeFilled as HomeIcon,
     AccountCircle as AccountIcon,
