+++ frontend/src/services/payment.service.ts
@@ -1,0 +1,6 @@
+import { request } from './http'
+import type { MidtransConfigResponse } from '../types'
+
+export function fetchMidtransClientKey() {
+  return request<MidtransConfigResponse>('/payments/config')
+}
