+++ frontend/src/types/category.type.ts
@@ -1,0 +1,6 @@
+export interface Category {
+  id: number
+  name: string
+  description?: string | null
+  icon?: string | null
+}
