+++ frontend/src/types/common.type.ts
@@ -1,0 +1,5 @@
+export interface SuccessResponse<T, S = null> {
+  message: string
+  data: T
+  metadata: S | null
+}
