+++ frontend/src/types/index.ts
@@ -1,0 +1,8 @@
+export * from './common.type'
+export * from './user.type'
+export * from './shop.type'
+export * from './category.type'
+export * from './product.type'
+export * from './order.type'
+export * from './payment.type'
+export * from './campaign.type'
