+++ frontend/src/types/shop.type.ts
@@ -1,0 +1,9 @@
+import type { User } from './user.type'
+
+export interface Shop {
+  id: number
+  name: string
+  description?: string | null
+  logo_url?: string | null
+  owner: User
+}
