+++ frontend/src/types/user.type.ts
@@ -1,0 +1,12 @@
+export interface User {
+  id: number
+  username: string
+  email: string
+}
+
+export interface Token {
+  access_token: string
+  token_type: string
+}
+
+export type AuthTokenResponse = Token
